requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pyyaml>=6.0.1
aiosqlite>=0.19.0
python-multipart>=0.0.6

# Proveedores de IA — instalar solo el que uses
anthropic>=0.40.0           # AI_PROVIDER=claude
openai>=1.0.0               # AI_PROVIDER=openai
google-generativeai>=0.8.0  # AI_PROVIDER=gemini

# Extraccion de texto de documentos (knowledge base)
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
